Replace Any::Moose with Moo
`Any::Moose` is deprecated in favour of `Moo` (see
https://metacpan.org/pod/Any::Moose and
http://shadow.cat/blog/matt-s-trout/moo-versus-any-moose/).  This commit
updates the code to use `Moo` syntax (which is in places slightly different
to `Moose`; e.g. `lazy_build` in `Moose` is `is lazy` in `Moo`) and updates
the object attributes appropriately.  The tests pass, so existing behaviour
has been retained.