Add initial Travis-CI configuration
It turns out that `Moose::Autobox` is an implicit dependency for the ROKR
`Dist::Zilla` plugin bundle and needed to be installed before the ROKR
bundle.

The `ReadmeFromPod` `Dist::Zilla` plugin was also found to be an implicit
dependency which wasn't later picked up by `dzil authordeps --missing` and
hence has been added to `dist.ini` in order to be installed by the above
command.

The dependencies install successfully and the test suite passes in this
configuration.